requirements.txt
absl-py==0.8.1
certifi==2019.11.28
Click==7.0
colorama==0.4.3
Cython==0.29.14
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.39
Jinja2==2.10.3
joblib==0.14.0
lxml==4.4.2
MarkupSafe==1.1.1
neo4j-driver==1.7.5
neobolt==1.7.15
neotime==1.7.4
numpy==1.17.3
openpyxl==3.0.2
pandas==0.25.3
Pillow==6.2.1
prompt-toolkit==2.0.10
py2neo==4.3.0
pyahocorasick==1.4.0
Pygments==2.3.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
rarfile==3.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
synonyms==3.10.2
urllib3==1.24.3
wcwidth==0.1.8
Werkzeug==0.16.0
Whoosh==2.7.4
wincertstore==0.2
